+++ src/main/java/edu/hm/hafner/analysis/parser/SonarQubeIssuesParser.java
@@ -9,8 +9,12 @@ import org.json.JSONObject;
  */
 public class SonarQubeIssuesParser extends SonarQubeParser {
     private static final long serialVersionUID = -8213765181968340929L;
-    
+
     private static final String ISSUE_SUB_PROJECT = "subProject";
+    private static final String ISSUE_TEXT_RANGE = "textRange";
+    private static final String ISSUE_TEXT_RANGE_START_LINE = "startLine";
+    private static final String ISSUE_TEXT_RANGE_END_LINE = "endLine";
+    private static final String ISSUE_LINE = "line";
 
     @Override
     protected boolean accepts(final JSONObject object) {
@@ -20,5 +24,21 @@ public class SonarQubeIssuesParser extends SonarQubeParser {
     @Override
     protected String getModulePath(final JSONObject component, final JSONObject issue) {
         return parseModulePath(issue, ISSUE_SUB_PROJECT);
+    }
+
+    @Override
+    protected int parseStart(final JSONObject issue) {
+        if (issue.has(ISSUE_TEXT_RANGE)) {
+            return issue.optJSONObject(ISSUE_TEXT_RANGE).optInt(ISSUE_TEXT_RANGE_START_LINE);
+        }
+        return issue.optInt(ISSUE_LINE);
+    }
+
+    @Override
+    protected int parseEnd(final JSONObject issue) {
+        if (issue.has(ISSUE_TEXT_RANGE)) {
+            return issue.optJSONObject(ISSUE_TEXT_RANGE).optInt(ISSUE_TEXT_RANGE_END_LINE);
+        }
+        return issue.optInt(ISSUE_LINE);
     }
 }